Added new validate command that validates a text layer json file
