+++ package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@google-cloud/vision": "^3.0.1",
+    "ajv": "^8.9.0",
     "python-shell": "^3.0.1",
     "rxjs": "^7.5.6",
     "ts-pattern": "^4.0.5",
